functions/vto-function/requirements.txt
google-genai>=0.8.0
flask>=2.0.0
google-cloud-secret-manager>=2.16.0              